vaadin project for the Graffiti
This project we created is embed with google map plugin. Also, we have
login and create account features integrated in the project. For each
user presented in the friend list, we can view the location of this user
by clicking the button associated with the user. By default, we also
have the features of zooming in and zooming out  for the purpose of
enlarge the certain view.